# phase-2/specs/001-todo-crud-features/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TodoList Pro API
  description: RESTful API for TodoList Pro application
  version: 1.0.0
  contact:
    name: TodoList Pro Team

servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.todolistpro.com/api/v1
    description: Production server

tags:
  - name: Authentication
    description: User registration, login, and session management
  - name: Todos
    description: Todo CRUD operations

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HTTP-only cookie with JWT token
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login with email and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HTTP-only cookie with JWT token
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout current user
      operationId: logoutUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Clears the auth cookie

  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user info
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /todos:
    get:
      tags:
        - Todos
      summary: List all todos for current user
      operationId: listTodos
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by completion status
          schema:
            type: string
            enum: [all, completed, remaining]
            default: all
        - name: search
          in: query
          description: Search in title and description (case-insensitive)
          schema:
            type: string
            maxLength: 100
        - name: cursor
          in: query
          description: Pagination cursor (created_at of last item)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Todos
      summary: Create a new todo
      operationId: createTodo
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoRead'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /todos/{todo_id}:
    get:
      tags:
        - Todos
      summary: Get a specific todo
      operationId: getTodo
      security:
        - cookieAuth: []
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoRead'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Todos
      summary: Update a todo
      operationId: updateTodo
      security:
        - cookieAuth: []
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoRead'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Todos
      summary: Delete a todo (soft delete)
      operationId: deleteTodo
      security:
        - cookieAuth: []
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /todos/{todo_id}/toggle:
    post:
      tags:
        - Todos
      summary: Toggle todo completion status
      operationId: toggleTodoStatus
      security:
        - cookieAuth: []
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoRead'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token

  schemas:
    UserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          example: "securepassword123"

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "securepassword123"

    UserRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserRead'
        message:
          type: string
          example: "Login successful"

    TodoCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: "Milk, eggs, bread"

    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy groceries (updated)"
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: "Milk, eggs, bread, butter"

    TodoRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        is_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    TodoListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TodoRead'
        next_cursor:
          type: string
          format: date-time
          nullable: true
          description: Cursor for next page, null if no more items
        total_count:
          type: integer
          description: Total number of items matching filter

    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: object
          properties:
            fields:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: "title"
                  error:
                    type: string
                    example: "Title must be between 1 and 200 characters"